+++ apps/omsorgspengerutbetaling-arbeidstaker-soknad/src/app/App.tsx
@@ -21,6 +21,7 @@ const {
     SIF_PUBLIC_APPSTATUS_DATASET,
     SIF_PUBLIC_APPSTATUS_PROJECT_ID,
     SIF_PUBLIC_USE_AMPLITUDE,
+    SIF_PUBLIC_AMPLITUDE_API_KEY,
     APP_VERSION,
 } = appEnv;
 
@@ -46,7 +47,8 @@ const App = () => (
             },
         }}
         publicPath={PUBLIC_PATH}
-        useAmplitude={SIF_PUBLIC_USE_AMPLITUDE ? SIF_PUBLIC_USE_AMPLITUDE === 'true' : isProd()}>
+        useAmplitude={SIF_PUBLIC_USE_AMPLITUDE ? SIF_PUBLIC_USE_AMPLITUDE === 'true' : isProd()}
+        amplitudeApiKey={SIF_PUBLIC_AMPLITUDE_API_KEY}>
         <SoknadApplicationCommonRoutes
             contentRoutes={[
                 <Route index key="redirect" element={<Navigate to={SøknadRoutes.VELKOMMEN} />} />,