+++ apps/omsorgspengerutbetaling-arbeidstaker-soknad/src/storybook/decorators/withAmplitudeProvider.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { AmplitudeProvider } from '@navikt/sif-common-amplitude';
 
 export const withAmplitudeProvider = (Story: any) => (
-    <AmplitudeProvider applicationKey={'ab'} isActive={false}>
+    <AmplitudeProvider applicationKey={'ab'} isActive={false} apiKey="default">
         <Story />
     </AmplitudeProvider>
 );