+++ apps/omsorgspengesoknad/src/storybook/decorators/withAmplitudeProvider.tsx
@@ -1,7 +1,7 @@
 import { AmplitudeProvider } from '@navikt/sif-common-amplitude';
 
 export const withAmplitudeProvider = (Story: any) => (
-    <AmplitudeProvider applicationKey={'ab'} isActive={false}>
+    <AmplitudeProvider applicationKey={'ab'} isActive={false} apiKey="default">
         <Story />
     </AmplitudeProvider>
 );