+++ apps/pleiepenger-sykt-barn/src/app/App.tsx
@@ -26,6 +26,7 @@ const {
     SIF_PUBLIC_APPSTATUS_PROJECT_ID,
     APP_VERSION,
     SIF_PUBLIC_USE_AMPLITUDE,
+    SIF_PUBLIC_AMPLITUDE_API_KEY,
 } = appEnv;
 ensureBaseNameForReactRouter(PUBLIC_PATH);
 
@@ -59,6 +60,7 @@ const App = () => {
             intlMessages={applicationIntlMessages}
             useAmplitude={SIF_PUBLIC_USE_AMPLITUDE ? SIF_PUBLIC_USE_AMPLITUDE === 'true' : isProd()}
             publicPath={PUBLIC_PATH}
+            amplitudeApiKey={SIF_PUBLIC_AMPLITUDE_API_KEY}
             onResetSoknad={handleResetSoknad}>
             <SoknadApplicationCommonRoutes
                 onReset={() => {