+++ apps/sif-ettersending/e2e/playwright/playwright.env.ts
@@ -5,6 +5,7 @@ export const playwrightEnv: AppEnv = {
     APP_VERSION: 'dev',
     PUBLIC_PATH: '/familie/sykdom-i-familien/soknad/ettersending',
     GITHUB_REF_NAME: 'branch-name',
+    SIF_PUBLIC_AMPLITUDE_API_KEY: 'default',
     SIF_PUBLIC_APPSTATUS_DATASET: 'staging',
     SIF_PUBLIC_APPSTATUS_PROJECT_ID: 'ryujtq87',
     SIF_PUBLIC_DEKORATOR_URL: