+++ apps/pleiepenger-sykt-barn/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @navikt/pleiepenger-sykt-barn
 
+## 4.4.3
+
+### Patch Changes
+
+-   48e9ecc: Sende med spesifikk apiKey for amplitude
+-   Updated dependencies [48e9ecc]
+    -   @navikt/sif-common-amplitude@2.26.22
+    -   @navikt/sif-common-soknad-ds@16.0.3
+    -   @navikt/sif-common-env@0.1.1
+    -   @navikt/sif-common-ui@0.8.28
+
 ## 4.4.2
 
 ### Patch Changes