+++ apps/ungdomsytelse-soknad/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @navikt/ungdomsytelse-søknad
 
+## 0.3.2
+
+### Patch Changes
+
+-   48e9ecc: Sende med spesifikk apiKey for amplitude
+-   Updated dependencies [48e9ecc]
+    -   @navikt/sif-common-soknad-ds@16.0.3
+    -   @navikt/sif-common-env@0.1.1
+    -   @navikt/sif-common-ui@0.8.28
+
 ## 0.3.1
 
 ### Patch Changes