+++ packages/sif-common-env/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @navikt/sif-common-ui
 
+## 0.1.1
+
+### Patch Changes
+
+-   48e9ecc: Sende med spesifikk apiKey for amplitude
+
 ## 0.1.0
 
 ### Minor Changes